Every component moved to separate file
