+++ myTrip/static/src/css/main.css
@@ -2,6 +2,12 @@
     height: 80px;
 }
 
+.title {
+    display:flex;
+    flex-wrap:nowrap;
+    align-items:center;
+}
+
 .title > a {
     color:white;
     text-decoration:none;
