+++ myTrip/static/src/notFound.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+export default class NotFound extends React.Component {
+    render() {
+        return <h1>Not found</h1>
+    }
+}
