+++ myTrip/static/src/registration/login.js
@@ -32,11 +32,15 @@ class LoginForm extends React.Component {
     handleSubmit(event) {
         const email = this.state.email;
         const password = this.state.password;
+        const EMAIL_REGEXP = /.+@.+\..+/;
         if (email == '') {
             this.setState({'emailError':'Email field is required'});
-            return false
+            return
+        } else if (email.match(EMAIL_REGEXP) == null) {
+            this.setState({'emailError':'Email fields is not in valid form'})
+            return
         } else {
-            this.setState({'emailError':''});
+            this.setState({'emailError':''})
         }
         if (password == '') {
             this.setState({'passwordError':'Password field is required'});
